requirements.txt
asgiref==3.7.2
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.74
cement==2.8.2
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.3
distlib==0.3.7
Django==4.2.6
django-simple-history==3.4.0
filelock==3.12.4
idna==3.4
jmespath==1.0.1
pathspec==0.10.1
pipenv==2023.9.8
platformdirs==3.10.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
urllib3==1.26.18
virtualenv==20.24.5
wcwidth==0.1.9
